Split r3159..r3161, part 12: Add explicit casts, non-pointer types.
Notes:
 - Atomic Edition (Censored) -> Plutonium Pak in grpscan.c
 - "scale" in astub.c:drawtileinfo() is incorrect, I think.
 - in demo.c, the gethitickms() value should be returned to a double.

git-svn-id: https://svn.eduke32.com/eduke32@3178 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/demo.c
@@ -728,7 +728,7 @@ nextdemo_nomenu:
 
                 if (Demo_IsProfiling())
                 {
-                    int32_t t = gethitickms();
+                    int32_t t = (int32_t)gethitickms();
                     G_DoMoveThings();
                     Demo_GToc(t);
                 }
